[PI-87] Minor refactoring
