+++ core/cardano-sl-core.cabal
@@ -42,6 +42,7 @@ library
                        Pos.Core.Class
                        Pos.Core.Common
                        Pos.Core.Configuration
+                       Pos.Core.ConfigPostgres
                        Pos.Core.Constants
                        Pos.Core.Context
                        Pos.Core.Delegation
