+++ crypto/cardano-sl-crypto.cabal
@@ -61,7 +61,6 @@ library
                      , lens
                      , memory
                      , mtl
-                     , postgresql-simple
                      , pvss
                      , quickcheck-instances
                      , reflection
